+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Introduce keywords, which are values of type binary, string or atom.
 - Use keywords for member names in object constraints.
 - Add a `values` constraint for strings.
+- Introduce definition catalogs.
 ## Misc
 - Rename the `element_type` constraint to `element`.